Problem: ROUTER socket discards messages over HWM
Use case is streaming of several thousand messages to a client at
startup. The ROUTER socket then drops messages as the HWM hits.

Solution: remove limit on queue. This will crash applications that
are badly designed. The proper solution is to use credit-based flow
control to limit traffic to any single client.