Problem: zproto server code logging is surprising
All other actors are using a command VERBOSE to enable logging;
the actor generated from zproto_server.gsl is using animation on
its state machine. This is surprising to external users, which
is not good.

Solution: use VERBOSE command and hide the fact that the server
has a state machine. Note: breaks the server config file, so
server/animate -> server/verbose.