Problem: caller should own message returned by _recv
Solution: do not destroy in generated code; allow caller to do this.

This is more consistent with what we'd expect from a CLASS API and
does not cost anything in performance terms. Also since the caller
will usually consume the message in order to process it, it also
makes sense.