Problem: client hands in destructor, on interrupt
As client destructor does a handshake with its background thread,
to allow a clean disconnect when possible, it hangs if the user
interrupted the process with Ctrl-C (as it never gets a reply
signal in s_accept_reply ()).

Solution: if zsys_interrupted is true, do not wait for a reply,
just return -1 immediately. The destructor then completes happily.