Problem: API destructor hangs if action = terminate is used
When using the "terminate" action, the state machine terminates
and then cannot handle a FLUSH request for the msgpipe. This means
the API destructor process hangs on flushing the msgpipe.

Solution: do not end the actor when the FSM calls "terminate",
rather just end the state machine, and keep processing the cmd and
msg pipes.