Problem: zproto client constructor does too much
The constructor is used to also connect to servers. This is neat except
that when it fails, the constructor fails, and there is no way to return
status information to the caller.

Solution: constructor should only initialize object. The connect method
should be done seperately.

- Removed support for "constructor" as special method in GSL code
- Removed call to constructor method, in class constructor

Note: this breaks existing client models and code. The fix is simple.
1. rename "constructor" to "connect" in XML model
2. in all calls to xxx_new, remove arguments
3. add second call to xxx_connect, with connection arguments
4. test return code 0/-1 from that method

See Malamute mlm_client as example.