Problem: codec can't decode from a buffer
As part of my attempt to write an integration test between Go, PHP
codecs and C codec, I figured we can't decode from and encode to
a buffer.

The idea is to store encoded buffer into a file so that the other
codec implementations can test their implementation against the
reference implementation (C codec).