Problem: need virtualized codec
This is to allow encoding/decoding into zmsg structure instead of
to/from socket, for carrying over Zyre, Malamute, and similar
protocols.

Solution: if class.virtual = 1, then send/recv are virtualized and
work with zmsg_t instead of zsock_t. Since we do not need both real
and virtual encoding in the same class, this keeps the API surface
to a minimum.

Note: I've not yet tested this except to be sure it hasn't broken
existing code.