An adaptation of the zproto_client_c.gsl generator that builds a zproto based Zyre node. This generator allows the creation of zyre nodes that use the zproto state machine scheme for generating edge-net  protocols. The generator wraps and interprets zyre's messages into the user's protocol space to be dealt with through the state machine syntax.
In addition to the generator I'm providing a zproject based example, loosely based on Zyre's chat example. I'm submitting this to the zproto project b.c. it felt like it might be a globally useful tool for other edge-net apps.

The generator and example are still a work in progress, I started building it as part of an edge-net based p2p game node. I will continue to update the generator as I develop the code further, I'm open to suggestions for making improvements as I go along.

 I wasn't sure the best place to put the files so they wouldn't clutter up the project. I can restructure it based on any recommendations.