Problem: Java String handling is incorrect for multi byte chars
The class uses the length of Strings instead of the length of the bytes
to calculate the frame size which causes the buffer to overflow when
multi byte chars are used.
It also does not use the ZMQ charset when getting bytes, thereby
switching between single byte and multi byte chars depending on the
underlying system.